+++ src/config.js
@@ -41,3 +41,18 @@ module.exports.loadConfig = () => {
     process.exit(1);
   }
 };
+
+const getModelBasedOnGengine = (engine) => {
+  if (process.env[`${engine.toUppercase()}_MODEL`]) {
+    return process.env[`${engine.toUppercase()}_MODEL`];
+  }
+  return process.env.MODEL;
+};
+
+module.exports.getSettingsByTemplate = (templateSettings, options) => {
+  const engine = options.engine || templateSettings.engine || process.env.ENGINE;
+  return {
+    engine,
+    model: options.engine || templateSettings.model || getModelBasedOnGengine(engine),
+  };
+};